feat: create role

+++ app/Http/Controllers/Api/Roles/RolesController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Role\CreateRoleRequest;
 use App\Http\Requests\Role\UpdateRoleRequest;
 use App\Models\Role\Role;
+use Illuminate\Support\Str;
 
 class RolesController extends Controller
 {
@@ -20,7 +21,22 @@ class RolesController extends Controller
     public function store(CreateRoleRequest $request)
     {
         $data = $request->validated();
-        $role = Role::create($data);
+
+        $data['slug'] = Str::slug($data['name']);
+
+        $permissions = $data['permissions'] ?? [];
+        unset($data['permissions']);
+        $role = Role::create([
+            'name' => $data['name'],
+            'slug' => $data['slug'],
+            'is_system_role' => false,
+            'description' => $data['description'] ?? null,
+        ]);
+
+        //todo relacionar as permissões
+
+        $role->permissions()->sync($permissions);
+
         return response()->json(['message' => 'Role created successfully', 'role' => $role], 201);
     }
 
@@ -28,6 +44,10 @@ class RolesController extends Controller
     public function update(UpdateRoleRequest $request, $id)
     {
         $role = Role::findOrFail($id);
+
+        if ($role->is_system_role) {
+            return response()->json(['message' => 'System roles cannot be updated'], 403);
+        }
 
         $data = $request->validated();
 
@@ -40,6 +60,10 @@ class RolesController extends Controller
     {
         //TODO autorizacao
         $role = Role::findOrFail($id);
+
+        if ($role->is_system_role) {
+            return response()->json(['message' => 'System roles cannot be deleted'], 403);
+        }
         $role->delete();
 
         return response()->json(['message' => 'Role deleted successfully'], 200);
